Make docker images of mysql and springboot api. Create Dockerfile to create image of springboot api and create data.sql file to create original user to ensure initial access to api

+++ java_backend_CRUD_1/src/main/java/crud_1/carSale/User.java
@@ -14,6 +14,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import lombok.Data;
@@ -26,7 +27,7 @@ public class User implements UserDetails{
 
 	@Id
 	@GeneratedValue
-	int id;
+	Long id;
 	
 	private String name;
 	private String password;
